KTL-4388: Extract KotlinToolingMetadataIndexingService from PackageIndexingService
