+++ app/src/main/kotlin/io/klibs/app/indexing/PackageIndexingService.kt
@@ -7,9 +7,6 @@ import io.klibs.core.pckg.dto.PackageDTO
 import io.klibs.core.pckg.entity.IndexingRequestEntity
 import io.klibs.core.pckg.enums.IndexingRequestStatus
 import io.klibs.core.pckg.enums.VersionType
-import io.klibs.core.pckg.model.Configuration
-import io.klibs.core.pckg.model.PackagePlatform
-import io.klibs.core.pckg.model.PackageTarget
 import io.klibs.core.pckg.repository.IndexingRequestRepository
 import io.klibs.core.pckg.repository.PackageRepository
 import io.klibs.core.pckg.service.PackageService
@@ -20,8 +17,6 @@ import io.klibs.integration.maven.MavenArtifact
 import io.klibs.integration.maven.MavenPom
 import io.klibs.integration.maven.MavenStaticDataProvider
 import io.klibs.integration.maven.delegate.KotlinToolingMetadataDelegate
-import io.klibs.integration.maven.delegate.KotlinToolingMetadataDelegateImpl
-import io.klibs.integration.maven.delegate.KotlinToolingMetadataDelegateStubImpl
 import jakarta.transaction.Transactional
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -33,7 +28,6 @@ import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.supervisorScope
-import org.jetbrains.kotlin.tooling.KotlinToolingMetadata
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.ObjectProvider
 import org.springframework.stereotype.Service
@@ -45,6 +39,7 @@ class PackageIndexingService(
     private val gitHubIndexingService: GitHubIndexingService,
     private val projectIndexingService: ProjectIndexingService,
     private val pomIndexingService: PomIndexingService,
+    private val kotlinToolingMetadataIndexingService: KotlinToolingMetadataIndexingService,
     private val packageDescriptionGenerator: PackageDescriptionGenerator,
 
     private val indexingRequestRepository: IndexingRequestRepository,
@@ -233,11 +228,10 @@ class PackageIndexingService(
             kotlinVersion = toolingMetadata.kotlinVersion,
             developers = pomIndexingService.extractDevelopers(pom),
             licenses = pomIndexingService.extractLicenses(pom),
-            configuration = toolingMetadata.toPackageConfiguration(),
+            configuration = kotlinToolingMetadataIndexingService.toPackageConfiguration(toolingMetadata),
             generatedDescription = descriptionWasGenerated,
             versionType = VersionType.from(pom.version),
-            targets = toolingMetadata.projectTargets.map { it.toPackageTarget() }
-                .distinct()
+            targets = kotlinToolingMetadataIndexingService.extractTargets(toolingMetadata)
         )
     }
 
@@ -265,130 +259,6 @@ class PackageIndexingService(
             }
         }
         return Pair(description, descriptionWasGenerated)
-    }
-
-    private fun KotlinToolingMetadataDelegate.toPackageConfiguration(): Configuration? {
-        return when (this) {
-            is KotlinToolingMetadataDelegateStubImpl -> null
-            is KotlinToolingMetadataDelegateImpl -> Configuration(
-                projectSettings = kotlinToolingMetadata.extractProjectSettings(),
-                jvmPlatform = kotlinToolingMetadata.extractJvmPlatformConfiguration(),
-                androidJvmPlatform = kotlinToolingMetadata.extractAndroidJvmPlatformConfiguration(),
-                nativePlatform = kotlinToolingMetadata.extractNativePlatformConfiguration(),
-                wasmPlatform = kotlinToolingMetadata.extractWasmPlatformConfiguration(),
-                jsPlatform = kotlinToolingMetadata.extractJsPlatformConfiguration(),
-            )
-        }
-    }
-
-    private fun KotlinToolingMetadata.extractProjectSettings(): Configuration.ProjectSettings {
-        return Configuration.ProjectSettings(
-            isHmppEnabled = this.projectSettings.isHmppEnabled,
-            isCompatibilityMetadataVariantEnabled = this.projectSettings.isCompatibilityMetadataVariantEnabled,
-        )
-    }
-
-    private fun KotlinToolingMetadata.extractJvmPlatformConfiguration(): Configuration.JvmPlatform? {
-        val jvmTarget = this.projectTargets.firstOrNull { it.platformType == "jvm" } ?: return null
-        val jvmExtras = jvmTarget.extras.jvm ?: return null
-        return Configuration.JvmPlatform(
-            jvmTarget = jvmExtras.jvmTarget,
-            withJavaEnabled = jvmExtras.withJavaEnabled
-        )
-    }
-
-    private fun KotlinToolingMetadata.extractAndroidJvmPlatformConfiguration(): Configuration.AndroidJvmPlatform? {
-        val androidJvmTarget = this.projectTargets.firstOrNull { it.platformType == "androidJvm" } ?: return null
-        val androidJvmExtras = androidJvmTarget.extras.android ?: return null
-        return Configuration.AndroidJvmPlatform(
-            sourceCompatibility = androidJvmExtras.sourceCompatibility,
-            targetCompatibility = androidJvmExtras.targetCompatibility
-        )
-    }
-
-    private fun KotlinToolingMetadata.extractNativePlatformConfiguration(): Configuration.NativePlatform? {
-        val nativeTargets = this.projectTargets.filter { it.platformType == "native" }
-        val chosenExtras = nativeTargets.firstNotNullOfOrNull { it.extras.native } ?: return null
-
-        val konanVersionsMatch = nativeTargets.all { isSameKonanVersion(it.extras.native, chosenExtras) }
-        require(konanVersionsMatch) {
-            "Konan configuration differs within one package: $this"
-        }
-
-        return Configuration.NativePlatform(
-            konanVersion = chosenExtras.konanVersion,
-            konanAbiVersion = chosenExtras.konanAbiVersion
-        )
-    }
-
-    private fun isSameKonanVersion(
-        first: KotlinToolingMetadata.ProjectTargetMetadata.NativeExtras?,
-        second: KotlinToolingMetadata.ProjectTargetMetadata.NativeExtras?
-    ): Boolean {
-        return first?.konanVersion == second?.konanVersion
-                && first?.konanAbiVersion == second?.konanAbiVersion
-    }
-
-    private fun KotlinToolingMetadata.extractWasmPlatformConfiguration(): Configuration.WasmPlatform? {
-        val wasmTarget = this.projectTargets.firstOrNull { it.platformType == "wasm" } ?: return null
-        val wasmExtras = wasmTarget.extras.js ?: return null
-        return Configuration.WasmPlatform(
-            isBrowserConfigured = wasmExtras.isBrowserConfigured,
-            isNodejsConfigured = wasmExtras.isNodejsConfigured
-        )
-    }
-
-    private fun KotlinToolingMetadata.extractJsPlatformConfiguration(): Configuration.JsPlatform? {
-        val jsTarget = this.projectTargets.firstOrNull { it.platformType == "js" } ?: return null
-        val jsExtras = jsTarget.extras.js ?: return null
-        return Configuration.JsPlatform(
-            isBrowserConfigured = jsExtras.isBrowserConfigured,
-            isNodejsConfigured = jsExtras.isNodejsConfigured
-        )
-    }
-
-    private fun KotlinToolingMetadata.ProjectTargetMetadata.toPackageTarget(): PackageTarget {
-        return PackageTarget(
-            platform = toPlatform(),
-            target = when (platformType) {
-                "common" -> null
-                "jvm" -> if (target == "com.android.build.api.variant.impl.KotlinMultiplatformAndroidLibraryTargetImpl") {
-                    // AGP 8.2-8.12
-                    extractAndroidTargetCompatibility()
-                } else {
-                    extras.jvm?.jvmTarget
-                }
-
-                "androidJvm" -> extras.android?.targetCompatibility
-                "wasm" -> null
-                "native" -> extras.native?.konanTarget
-                "js" -> null
-                else -> error("Unknown platform type: $platformType")
-            }
-        )
-    }
-
-    private fun extractAndroidTargetCompatibility(): String {
-        // Safe option, too hard to extract actual compatibility
-        return "1.8"
-    }
-
-    private fun KotlinToolingMetadata.ProjectTargetMetadata.toPlatform(): PackagePlatform {
-        return when (this.platformType) {
-            "common" -> PackagePlatform.COMMON
-            "jvm" -> if (target == "com.android.build.api.variant.impl.KotlinMultiplatformAndroidLibraryTargetImpl") {
-                // AGP 8.2-8.12
-                PackagePlatform.ANDROIDJVM
-            } else {
-                PackagePlatform.JVM
-            }
-
-            "androidJvm" -> PackagePlatform.ANDROIDJVM
-            "wasm" -> PackagePlatform.WASM
-            "native" -> PackagePlatform.NATIVE
-            "js" -> PackagePlatform.JS
-            else -> error("Unknown platform type: ${this.platformType}")
-        }
     }
 
     private companion object {
